Create Navbar component. Use forwardRef to access homeclick props from homepage component. Add iconType to types file
